Update UserInterface.py
Functional as of now. It has a status indicator for the status of connecting the Arduino, it reads live data, displays the target angle and set speed, is able to TARE although that needs to be looked over. also displays direction.

TODO: fix location of buttons and variables. and include live scrolling for table and table